Refactor custom patterns into modules
